Move blinding implementation, eventReaderRoot to AnalysisFramework.
Initially I put the blinding implentation in eventReaderRoot as I
wanted to manipulate the events under the nose of the analyzer. But
the more I thought about it, it seemed wrong. The blinding isn't
really part of the data. It's part of the analysis.

Also by inheriting from Cosmin's AnitaDataset class I don't need to
distribute blind header files, I can have a small tree of manipulated
headers like I do with the UsefulAnitaEvents. Since both the header
and the event go through the AnitaDataset class, both can be
manipulated at once and swapped out seemlessly.