Revert Double32_t->Double_t, bug fix trainingPeak, update fracPower
Undo Cosmin's Double32_t changes after meeting discussion.
New algorithm for finding the training peak since it was producing the
wrap answer a certain fraction of the time for WAIS and was probably
more wrong for MC.
Rename narrowestWidths to more descriptive fracPowerWinow and store
beginning and end instead of just window width.